requirements/base.txt
beautifulsoup4==4.4.0
Django==1.8.4
Markdown==2.6.2
Pillow==2.9.0
# South ## gli: no longer in django
cmsplugin-filer==0.10.2
cmsplugin-news==0.4.4
django-braces==1.8.1
django-classy-tags==0.6.2
django-cms==3.1.3
djangocms-inherit==0.1 ## gli: for djangocms 3.0
djangocms-googlemap==0.3 ## gli: for djangocms 3.0
djangocms-link==1.6.2 ## gli: for djangocms 3.0
djangocms-snippet==1.5 ## gli: for djangocms 3.0
django-crispy-forms==1.5.1
django-compressor==1.5
django-filer==0.9.12
django-gravatar2==1.3.0
# django-timezones ## gli: not compatible with django 1.8
-e git+https://github.com/jezdez/django-configurations.git@0ff4c7612d28036449edd4b2cf3b58ca1eded0f2#egg=django_configurations-master
dj-database-url==0.3.0
django-statici18n==1.1.5
django-validatorchain==0.1
django-sortedm2m==1.0.2

# Strict due to API compatibility issues
django-treebeard==3.0 ## gli: for djangocms 3.1

django-reversion==1.9.3
django-sekizai==0.8.2
django-social-auth3==0.7.23
django-taggit==0.17.0
djangocms-admin-style==0.2.7
djangocms-text-ckeditor==2.6.0
djangocms-style==1.5
-e git+https://github.com/zerok/django-userprofiles.git@fe29d7feb2a1b0ca8a10be8032f235df5406267f#egg=django_userprofiles-origin_pullrequests_emailverification-manager
easy-thumbnails==2.2
html5lib==0.999999
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
gunicorn==19.3.0
psycopg2==2.6.1

lxml==3.4.4

# tablib for export support
tablib==0.10.0

# Search integration
django-haystack==2.4.0

-e git+https://github.com/kliment/pymill.git@f55cab94826168765a5a7d60ffed2641fe89b27d#egg=pymill-dev

# For now we also include Whoosh here. In the future this will
# be moved to dev.txt
Whoosh==2.7.0

# Redis as caching backend
django-redis==4.2.0
hiredis==0.2.0
redis==2.10.3

# Celery for async tasks (pdf generation)
celery==3.1.18

# Documentation
Sphinx==1.3.1
sphinx-rtd-theme==0.1.8
